+++ internal/discover/service/plugins/ard.go
@@ -10,6 +10,7 @@ import (
 	"github.com/Method-Security/networkscan/generated/go/common"
 	"github.com/Method-Security/networkscan/generated/go/common/protocol"
 	discoverfern "github.com/Method-Security/networkscan/generated/go/discover"
+	"github.com/Method-Security/networkscan/internal/discover/service/helpers"
 )
 
 type ArdFingerprinter struct{}
@@ -20,14 +21,14 @@ func (ArdFingerprinter) DefaultPorts() []int { return []int{3283, 5900} }
 
 func (ArdFingerprinter) Detect(ctx context.Context, ip net.IP, port int, host string, timeout int) (*discoverfern.ServiceDetails, error) {
 	addr := net.JoinHostPort(ip.String(), fmt.Sprintf("%d", port))
-	conn, err := dialService(ctx, "tcp", addr, timeout)
+	conn, err := helpers.Dial(ctx, "tcp", addr, timeout)
 	if err != nil {
 		return nil, err
 	}
 	defer func() { _ = conn.Close() }()
 
 	// Set read/write deadline
-	if err := setServiceDeadline(conn, timeout); err != nil {
+	if err := helpers.SetDeadline(conn, timeout); err != nil {
 		return nil, err
 	}
