+++ internal/discover/service/plugins/dcerpc.go
@@ -10,6 +10,7 @@ import (
 	"github.com/Method-Security/networkscan/generated/go/common"
 	"github.com/Method-Security/networkscan/generated/go/common/protocol"
 	discoverfern "github.com/Method-Security/networkscan/generated/go/discover"
+	"github.com/Method-Security/networkscan/internal/discover/service/helpers"
 	"github.com/Method-Security/networkscan/utils"
 )
 
@@ -20,7 +21,7 @@ func (DCERPCFingerprinter) Name() string { return "dcerpc" }
 func (DCERPCFingerprinter) DefaultPorts() []int { return []int{135} }
 
 func (DCERPCFingerprinter) Detect(ctx context.Context, ip net.IP, port int, host string, timeout int) (*discoverfern.ServiceDetails, error) {
-	conn, err := dialService(ctx, "tcp", utils.FormatHostPort(ip.String(), port), timeout)
+	conn, err := helpers.Dial(ctx, "tcp", utils.FormatHostPort(ip.String(), port), timeout)
 	if err != nil {
 		return nil, err
 	}
@@ -30,7 +31,7 @@ func (DCERPCFingerprinter) Detect(ctx context.Context, ip net.IP, port int, host
 	bindRequest := buildDCERPCBindRequest()
 
 	// Set deadline
-	if err := setServiceDeadline(conn, timeout); err != nil {
+	if err := helpers.SetDeadline(conn, timeout); err != nil {
 		return nil, err
 	}
