+++ internal/discover/service/plugins/dhcp.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/Method-Security/networkscan/generated/go/common"
 	discoverfern "github.com/Method-Security/networkscan/generated/go/discover"
+	"github.com/Method-Security/networkscan/internal/discover/service/helpers"
 )
 
 type DHCPFingerprinter struct{}
@@ -21,14 +22,14 @@ func (DHCPFingerprinter) Detect(ctx context.Context, ip net.IP, port int, host s
 	addr := net.JoinHostPort(ip.String(), fmt.Sprintf("%d", port))
 
 	// Create UDP connection
-	conn, err := dialService(ctx, "udp", addr, timeout)
+	conn, err := helpers.Dial(ctx, "udp", addr, timeout)
 	if err != nil {
 		return nil, err
 	}
 	defer func() { _ = conn.Close() }()
 
 	// Set read deadline
-	if err := setServiceReadDeadline(conn, timeout); err != nil {
+	if err := helpers.SetReadDeadline(conn, timeout); err != nil {
 		return nil, err
 	}
