+++ internal/discover/service/plugins/ldap.go
@@ -10,6 +10,7 @@ import (
 	"github.com/Method-Security/networkscan/generated/go/common"
 	"github.com/Method-Security/networkscan/generated/go/common/protocol"
 	discoverfern "github.com/Method-Security/networkscan/generated/go/discover"
+	"github.com/Method-Security/networkscan/internal/discover/service/helpers"
 	"github.com/go-ldap/ldap/v3"
 )
 
@@ -19,14 +20,14 @@ func (LDAPFingerprinter) Name() string { return "ldap" }
 
 func (LDAPFingerprinter) Detect(ctx context.Context, ip net.IP, port int, host string, timeout int) (*discoverfern.ServiceDetails, error) {
 	// Create a context using the scanner timeout.
-	_, cancel := serviceContext(ctx, timeout)
+	_, cancel := helpers.Context(ctx, timeout)
 	defer cancel()
 
 	// Use proper LDAP client with timeout
 	addr := net.JoinHostPort(ip.String(), fmt.Sprintf("%d", port))
 
 	conn, err := ldap.DialURL(fmt.Sprintf("ldap://%s", addr), ldap.DialWithDialer(&net.Dialer{
-		Timeout: serviceTimeout(timeout),
+		Timeout: helpers.Timeout(timeout),
 	}))
 	if err != nil {
 		return nil, err
