+++ internal/discover/service/plugins/mongodb.go
@@ -9,6 +9,7 @@ import (
 	"github.com/Method-Security/networkscan/generated/go/common"
 	"github.com/Method-Security/networkscan/generated/go/common/protocol"
 	discoverfern "github.com/Method-Security/networkscan/generated/go/discover"
+	"github.com/Method-Security/networkscan/internal/discover/service/helpers"
 	"github.com/Method-Security/networkscan/utils"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
@@ -30,8 +31,8 @@ func (MongoDBFingerprinter) Detect(ctx context.Context, ip net.IP, port int, hos
 	uri := fmt.Sprintf("mongodb://%s", addr)
 
 	// Create context with timeout
-	timeoutDuration := serviceTimeout(timeout)
-	timeoutCtx, cancel := serviceContext(ctx, timeout)
+	timeoutDuration := helpers.Timeout(timeout)
+	timeoutCtx, cancel := helpers.Context(ctx, timeout)
 	defer cancel()
 
 	// Set client options with short timeout
