+++ internal/discover/service/plugins/smb.go
@@ -8,6 +8,7 @@ import (
 	"github.com/Method-Security/networkscan/generated/go/common"
 	commonprotocolfern "github.com/Method-Security/networkscan/generated/go/common/protocol"
 	discoverfern "github.com/Method-Security/networkscan/generated/go/discover"
+	"github.com/Method-Security/networkscan/internal/discover/service/helpers"
 	smbclient "github.com/Method-Security/networkscan/internal/protocol/smb"
 )
 
@@ -44,7 +45,7 @@ func (SMBFingerprinter) Detect(ctx context.Context, ip net.IP, port int, host st
 		return result, nil
 	case err := <-errorChan:
 		return nil, err
-	case <-serviceAfter(timeout):
+	case <-helpers.After(timeout):
 		return nil, context.DeadlineExceeded
 	case <-ctx.Done():
 		return nil, ctx.Err()
