+++ internal/discover/service/plugins/winrm.go
@@ -13,6 +13,7 @@ import (
 	"github.com/Method-Security/networkscan/generated/go/common"
 	"github.com/Method-Security/networkscan/generated/go/common/protocol"
 	discoverfern "github.com/Method-Security/networkscan/generated/go/discover"
+	"github.com/Method-Security/networkscan/internal/discover/service/helpers"
 )
 
 type WinRMFingerprinter struct{}
@@ -53,7 +54,7 @@ func detectWinRMWithScheme(ctx context.Context, ip net.IP, port int, host string
 
 	// Create HTTP client
 	client := &http.Client{
-		Timeout: serviceTimeout(timeout),
+		Timeout: helpers.Timeout(timeout),
 		Transport: &http.Transport{
 			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
 		},
